RAG_Chatbot/backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-groq
langchain-huggingface
sentence-transformers
chromadb==0.4.24
pandas==2.2.1
python-dotenv==1.0.1
unstructured==0.12.5
markdown==3.5.2
